requirements.txt
torch
transformers
# transformers@git+https://github.com/rimon15/transformers
tqdm
einops
ruff
jaxtyping
beartype
pre-commit
jupyter
matplotlib
scikit-learn
scipy
umap-learn
accelerate
datasets
evil_twins@git+https://github.com/rimon15/evil_twins.git
nanogcg
nnsight
python-dotenv
boto3
awscli
skypilot[aws]
seaborn
spacy
lingua-language-detector
sentencepiece
termcolor
tabulate
pandas
pybind11
omegaconf
wandb
nltk
smplotlib==0.0.9